requirements.txt
streamlit
langchain
langchain-google-genai
pandas
python-dotenv
langchain-experimental
tabulate
